+++ faye/src/components/Main.jsx
@@ -1,28 +1,11 @@
 import React from "react";
 import "../App.css";
 import { useState } from "react";
+import Header from "./Header";
+import LeftPage from "./LeftPage";
+import RightPage from "./RightPage";
 
 function Main() {
-  const [count, setCount] = useState(0);
-  const [currentDate, setCurrentDate] = useState(new Date());
-
-  const options = {
-    weekday: "short",
-    year: "numeric",
-    month: "2-digit",
-    day: "2-digit",
-  };
-  const formattedDate = currentDate
-    .toLocaleDateString(undefined, options)
-    .replace(/\s/g, "")
-    .replace(/-/g, "/")
-    .replace(/,/g, "/");
-
-  const weekdayIndex = formattedDate.indexOf("/");
-  const formattedDateUpperCase =
-    formattedDate.slice(0, weekdayIndex).toUpperCase() +
-    formattedDate.slice(weekdayIndex);
-
   return (
     <div>
       {" "}
@@ -45,107 +28,12 @@ function Main() {
               {/* Main Div/Container */}
               <div className="mx-4 w-full bg-    relative z-10 rounded-lg">
                 {/* Header/Nav */}
-                <div className="h-8 mt-4 w-full flex justify-center items-center y2k-shadow4 bg-gradient-to-r from-[#cee3e5] to-[#b4cdcf] rounded-lg inner-stroke2 opacity-90 ">
-                  <div className="flex justify-end  w-full">
-                    <h1 className="rounded-md   bg-white y2k-shadow4 mx-4 w-1/6   ">
-                      {formattedDateUpperCase}
-                    </h1>
-                  </div>
-                </div>
+                <Header />
                 <div className="flex h-[48rem] p-4 ">
                   {/* Left Side */}
-                  <div className="w-2/3 flex flex-col">
-                    <div className="h-5/6 bg-blue-300 mx-1 mt-2 rounded-xl y2k-shadow2 flex justify-center items-center ">
-                      <div className="h-full rounded-2xl  w-full bg-cover y2k-shadow4 bg-[url('./assets/faye.png')] inner-stroke "></div>
-                    </div>
-                    <div className="h-1/6 bg-gradient-to-t from-[#8ea7aa] to-[#53686b] mb-1 mt-4 rounded-2xl y2k-shadow4 mx-1 p-4 inner-stroke">
-                      <div className="flex items-center justify-center h-full">
-                        <div className="w-1/2 flex flex-col items-start justify-center pr-2">
-                          <div className="w-full bg-gray-400 m-1 y2k-shadow3 flex">
-                            <div className="text-right p-1 text-white bg-green-500 w-1/3 y2k-shadow3"></div>
-                          </div>
-                          <div className="w-full bg-gray-400 m-1 y2k-shadow3 flex">
-                            <div className="text-right p-1 text-white bg-green-500 w-1/2 y2k-shadow3"></div>
-                          </div>
-                          <div className="w-full bg-gray-400 m-1 y2k-shadow3 flex">
-                            <div className="text-right p-1 text-white bg-green-500 w-1/2 y2k-shadow3"></div>
-                          </div>
-                          <div className="w-full bg-gray-400 m-1 y2k-shadow3 flex">
-                            <div className="text-right p-1 text-white bg-green-500 w-1/3 y2k-shadow3"></div>
-                          </div>
-                          <div className="w-full bg-gray-400 m-1 y2k-shadow3 flex">
-                            <div className="text-right p-1 text-white bg-green-500 w-1/2 y2k-shadow3"></div>
-                          </div>
-                          <div className="w-full bg-gray-400 m-1 y2k-shadow3 flex">
-                            <div className="text-right p-1 text-white bg-green-500 w-1/2 y2k-shadow3"></div>
-                          </div>
-                        </div>
-                        <div className="w-1/2 flex flex-col items-start justify-center pl-2">
-                          <div className="w-full  m-1  flex">
-                            <div className="text-right p-1 text-white bg-green-500 w-1/3 y2k-shadow3"></div>
-                          </div>
-                          <div className="w-full  m-1  flex">
-                            <div className="text-right p-1 text-white bg-green-500 w-1/2 y2k-shadow3"></div>
-                          </div>
-                          <div className="w-full  m-1  flex">
-                            <div className="text-right p-1 text-white bg-green-500 w-1/2 y2k-shadow3"></div>
-                          </div>
-                          <div className="w-full  m-1  flex">
-                            <div className="text-right p-1 text-white bg-green-500 w-1/3 y2k-shadow3"></div>
-                          </div>
-                          <div className="w-full  m-1  flex">
-                            <div className="text-right p-1 text-white bg-green-500 w-1/2 y2k-shadow3"></div>
-                          </div>
-                          <div className="w-full  m-1  flex">
-                            <div className="text-right p-1 text-white bg-green-500 w-1/2 y2k-shadow3"></div>
-                          </div>
-                        </div>
-                      </div>
-                    </div>
-                  </div>
+                  <LeftPage />
                   {/* Right Side */}
-                  <div className="w-1/2 flex flex-col ">
-                    <div className="h-1/6 bg-gradient-to-t from-[#cee3e5] to-[#b4cdcf] mx-2 mt-2 rounded-md y2k-shadow4 inner-stroke flex justify-evenly items-center">
-                      {/* Left Arrow */}
-                      <div className="xp-button-container text-gray-400">
-                        <button className="" disabled>
-                          Back
-                        </button>
-                      </div>
-                      1/3
-                      <div className="xp-button-container ">
-                        <button className="">Forward</button>
-                      </div>
-                    </div>
-
-                    <div className="h-2/6 bg-gradient-to-t from-[#000000] to-[#000000] text-white mx-2 mt-2 rounded-md y2k-shadow4 inner-stroke">
-                      <div class="window-body">
-                        <div class="title-bar">
-                          <div class="title-bar-text">About</div>
-                          <div class="title-bar-controls"></div>
-                        </div>
-                        <div className="flex mx-4 mt-2 ">
-                          Final Fantasy X[a] is a role-playing video game
-                          developed and published by Square as the tenth main
-                          entry in the Final Fantasy series. Originally released
-                          in 2001 for PlayStation 2.
-                        </div>
-                      </div>
-                    </div>
-                    <div className="h-3/6 flex mx-2 mt-2 ">
-                      <div className="h-full w-full flex flex-col">
-                        <div className="h-full bg-gradient-to-t from-[#cee3e5] to-[#b4cdcf] mb- rounded-md y2k-shadow4 inner-stroke"></div>
-                        <div className="h-4/6 bg-gradient-to-t from-[#8ea7aa] to-[#53686b] y2k-shadow3 rounded-md mt-2 mb-1 inner-stroke"></div>
-                      </div>
-                      <div className="w-1/2 h-full bg-gradient-to-t from-[#4e6365] to-[#b5d0d2] ml-2 rounded-md flex flex-col items-center inner-stroke">
-                        <div className="h-6 mt-2 w-5/6 bg-[#8ba9b0] rounded-md y2k-shadow4 border-y2k"></div>
-                        <div className="h-4 mt-2 w-1/2 bg-[#8ba9b0] rounded-md y2k-shadow4 border-y2k"></div>
-                        <div className="h-1/3 w-2/3 mt-4 bg-[#818989] y2k-shadow2 rounded-md "></div>
-                        <div className="h-1/6 w-2/3 mt-4 bg-[#818989] y2k-shadow2 rounded-md "></div>
-                        <div className="h-1/6 w-2/3 mt-4 bg-[#818989] y2k-shadow2 rounded-md "></div>
-                      </div>
-                    </div>
-                  </div>
+                  <RightPage />
                 </div>
               </div>
             </div>
